% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_connect.R
\name{vs_connect}
\alias{vs_connect}
\title{Creating a variantspark connection}
\usage{
vs_connect(sc)
}
\arguments{
\item{sc}{A spark connection.}
}
\value{
A variantspark connection
}
\description{
You need to create a variantspark connection to use this extension. To do this,
you pass as argument a spark connection that you can create
using \code{sparklyr::spark_connect()}.
}
\examples{
library(sparklyr)
sc <- spark_connect(master = "spark://HOST:PORT")
connection_is_open(sc)
vsc <- vs_connect(sc)
spark_disconnect(sc)

}
