% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_tbl.R
\name{importance_tbl}
\alias{importance_tbl}
\title{Extract the importance data frame}
\usage{
importance_tbl(importance, name = "importance_tbl")
}
\arguments{
\item{importance}{A jobj from the class \code{ImportanceAnalysis}, usually the
output of \code{vs_importance_analysis()}.}

\item{name}{The name to assign to the copied table in Spark.}
}
\description{
This function extracts the importance data frame from the Importance Analysis
jobj.
}
